+++ lib/sidekiq/redis_connection.rb
@@ -6,16 +6,17 @@ module Sidekiq
   class RedisConnection
     def self.create(options={})
       url = options[:url] || ENV['REDISTOGO_URL'] || 'redis://localhost:6379/0'
+      driver = options[:driver] || 'ruby'
       # need a connection for Fetcher and Retry
       size = options[:size] || (Sidekiq.server? ? (Sidekiq.options[:concurrency] + 2) : 5)
 
       ConnectionPool.new(:timeout => 1, :size => size) do
-        build_client(url, options[:namespace])
+        build_client(url, options[:namespace], driver)
       end
     end
 
-    def self.build_client(url, namespace)
-      client = Redis.connect(:url => url)
+    def self.build_client(url, namespace, driver)
+      client = Redis.connect(:url => url, :driver => driver)
       if namespace
         Redis::Namespace.new(namespace, :redis => client)
       else