+++ test/test_client.rb
@@ -3,36 +3,6 @@ require 'sidekiq/client'
 require 'sidekiq/worker'
 
 class TestClient < MiniTest::Unit::TestCase
-  describe 'with real redis' do
-    before do
-      Sidekiq.redis = REDIS
-      Sidekiq.redis {|c| c.flushdb }
-    end
-
-    class QueueWorker
-      include Sidekiq::Worker
-      sidekiq_options :queue => 'customqueue'
-    end
-
-    it 'does not push duplicate messages when configured for unique only' do
-      Sidekiq.client_middleware do |chain|
-        chain.add Sidekiq::Middleware::Client::UniqueJobs
-      end
-      QueueWorker.sidekiq_options :unique => true
-      10.times { Sidekiq::Client.push('class' => QueueWorker, 'args' => [1, 2]) }
-      assert_equal 1, Sidekiq.redis {|c| c.llen("queue:customqueue") }
-    end
-
-    it 'does push duplicate messages when not configured for unique only' do
-      Sidekiq.client_middleware do |chain|
-        chain.add Sidekiq::Middleware::Client::UniqueJobs
-      end
-      QueueWorker.sidekiq_options :unique => false
-      10.times { Sidekiq::Client.push('class' => QueueWorker, 'args' => [1, 2]) }
-      assert_equal 10, Sidekiq.redis {|c| c.llen("queue:customqueue") }
-    end
-  end
-
   describe 'with mock redis' do
     before do
       @redis = MiniTest::Mock.new