+++ test/test_middleware.rb
@@ -1,6 +1,5 @@
 require 'helper'
 require 'sidekiq/middleware/chain'
-require 'sidekiq/middleware/server/unique_jobs'
 require 'sidekiq/processor'
 
 class TestMiddleware < MiniTest::Unit::TestCase
@@ -8,18 +7,6 @@ class TestMiddleware < MiniTest::Unit::TestCase
     before do
       $errors = []
       Sidekiq.redis = REDIS
-    end
-
-    it 'handles errors' do
-      handler = Sidekiq::Middleware::Server::ExceptionHandler.new
-
-      assert_raises ArgumentError do
-        handler.call('', { :a => 1 }, 'default') do
-          raise ArgumentError
-        end
-      end
-      assert_equal 1, $errors.size
-      assert_equal({ :a => 1 }, $errors[0][:parameters])
     end
 
     class CustomMiddleware
@@ -84,10 +71,3 @@ class TestMiddleware < MiniTest::Unit::TestCase
     end
   end
 end
-
-class FakeAirbrake
-  def self.notify(ex, hash)
-    $errors << hash
-  end
-end
-Airbrake = FakeAirbrake